parte_2/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
supabase==2.0.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0